loader/app/requirements.txt
azure-identity
azure-storage-blob
pandas
pymongo
azure-identity